WebDevelopmentLab03/requirements.txt
streamlit
google-generativeai
request